Expose KeyEvents mixin
