+++ app/mixins/key-events.js
@@ -1,27 +1,4 @@
 import Ember from 'ember';
+import KeyEventsMixin from 'ember-cli-paint/mixins/key-events';
 
-const keyCodeToEventMap = {
-  27: 'esc',
-  37: 'leftArrow',
-  39: 'rightArrow'
-};
-
-export default Ember.Mixin.create({
-  mergedProperties: ['keyEvents'],
-  keyEvents: {},
-
-  setupKeyHandling: function() {
-    this.$(document).on(`keyup.${this.get('elementId')}`, (event) => {
-      var key = keyCodeToEventMap[event.keyCode];
-      var keyEvent = this.get('keyEvents')[key];
-
-      if (keyEvent) {
-        keyEvent.apply(this);
-      }
-    });
-  }.on('didInsertElement'),
-
-  tearDownKeyHandling: function() {
-    this.$(document).off(`keyup.${this.get('elementId')}`);
-  }.on('willDestroyElement')
-});
+export default KeyEventsMixin;
